+++ R/plot_brickchart.R
@@ -3,14 +3,16 @@
 #' Plot bars of proportions that consist of "bricks" showing individual
 #' observations.
 #'
-#' @param data Data set
-#' @param outcome Outcome expression, e.g., \code{event == TRUE}
-#' @param by Exposure variable
-#' @param group Optional: Grouping variable, e.g. an effect modifier.
-#' @param colors Color list. Must be a \code{list} consisting of two-element
-#'   color code vectors with the dark and bright colors. The default
-#'   provides colors for three exposure groups; additions may be necessary.
+#' @param data Data set.
+#' @param outcome Outcome expression, e.g., \code{event == TRUE}.
+#' @param by Exposure variable.
+#' @param group Optional: Grouping variable, e.g., an effect modifier.
+#' @param colors Optional: Color list. Must be a \code{list} consisting of
+#'   two-element color code vectors with the dark and bright colors for each
+#'   level of the exposure variable (\code{by}).
 #'   Example: \code{list(c("darkred", "red"), c("darkblue", "lightblue"))}.
+#'   If not provided, colors will be generated from the
+#'   \code{\link[viridis]{viridis_pal}} palette.
 #' @param guide Optional: Show legend? Defaults to \code{FALSE}.
 #'
 #' @return ggplot. Modify further with standard ggplot functions.
@@ -32,29 +34,44 @@
 #'   dplyr::filter(ph.ecog < 3) %>%
 #'   brickchart(outcome = status == 2,
 #'              by = ph.ecog,
-#'              group = sex)
+#'              group = sex,
+#'              guide = TRUE) +  # show color legend
+#'   # Modify graph with standard ggplot functions
+#'   # Refer to axes before flipping x <-> y. Here, y is horizontal:
+#'   ggplot2::labs(y = "Risk (cumulative incidence)",
+#'                 fill = "ECOG status",  # Color label
+#'                 title = "Mortality by ECOG performance status") +
+#'   # Themes refer to axes as shown--'y' is now vertical
+#'   ggplot2::theme(axis.title.y = ggplot2::element_blank())
 brickchart <- function(
   data, outcome, by, group,
-  colors = list(
-    c(viridis::viridis_pal(end = 0.9, option = "cividis", alpha = 1)(3)[1],
-      viridis::viridis_pal(end = 0.9, option = "cividis", alpha = 0.8)(3)[1]),
-    c(viridis::viridis_pal(end = 0.9, option = "cividis", alpha = 1)(3)[2],
-      viridis::viridis_pal(end = 0.9, option = "cividis", alpha = 0.8)(3)[2]),
-    c(viridis::viridis_pal(end = 0.9, option = "cividis", alpha = 1)(3)[3],
-      viridis::viridis_pal(end = 0.9, option = "cividis", alpha = 0.8)(3)[3])),
+  colors = NULL,
   guide = FALSE) {
-  data <- data %>% dplyr::mutate({{ by }} := forcats::fct_rev(factor({{ by }})))
-  if(missing(group))
+  data <- data %>%
+    dplyr::mutate({{ by }} := forcats::fct_rev(factor({{ by }})))
+  if(missing(group)) {
     group <- NULL  # for facet_grid
-  else
+  } else {
     data <- data %>% dplyr::arrange({{ group }})
+  }
+  if(is.null(colors)) {
+    by_length <- length(unique(data %>% dplyr::pull({{ by }})))
+    colors <- purrr::map(
+      .x = 1:by_length,
+      .f = ~c(viridis::viridis_pal(end = 0.9,
+                                   option = "cividis",
+                                   alpha = 1)(by_length)[.x],
+              viridis::viridis_pal(end = 0.9,
+                                   option = "cividis",
+                                   alpha = 0.8)(by_length)[.x]))
+  }
   fillcolors <- data %>%
     dplyr::filter({{ outcome }}) %>%
     dplyr::count({{ by }}) %>%
     dplyr::mutate(index = dplyr::row_number()) %>%
     dplyr::left_join(tibble::tibble(colors = colors) %>%
                        dplyr::mutate(index = dplyr::row_number()),
-              by = "index") %>%
+                     by = "index") %>%
     dplyr::mutate(colors = purrr::map2(.x = colors, .y = n,
                                        .f = ~rep(.x, length.out = .y))) %>%
     dplyr::pull(colors) %>%