+++ R/plots.R
@@ -396,7 +396,7 @@ exclusion_flowchart <- function(design, width = 4) {
 #' # an RMST difference, having drawn the plot via
 #' #   figure <- survminer::ggsurvplot(...)
 #' \dontrun{
-#' df_ribbon <- figure$plot$data %>%  # acces plot data
+#' df_ribbon <- figure$plot$data %>%  # access plot data
 #'   select(time, surv, strata) %>%
 #'   pivot_wider(names_from = strata,
 #'               values_from = surv,